+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 
 Nothing unreleased.
 
+## [6.1.0] - 2025-03-29
+
+- Added microsecond precision to UUIDv7;
+- Added methods to `GUID`;
+- Refactored many classes;
+- Updated some tests.
+
 ## [6.0.0] - 2024-07-07
 
 This version has breaking changes.